+++ ethcore/src/client/mod.rs
@@ -18,13 +18,12 @@
 
 mod client;
 mod config;
-mod ids;
 mod test_client;
 mod trace;
 
 pub use self::client::*;
 pub use self::config::{ClientConfig, BlockQueueConfig, BlockChainConfig, Switch};
-pub use self::ids::{BlockId, TransactionId, UncleId, TraceId};
+pub use types::ids::*;
 pub use self::test_client::{TestBlockChainClient, EachBlockWith};
 pub use self::trace::Filter as TraceFilter;
 pub use executive::{Executed, Executive, TransactOptions};
